create the Nav component (previously a part of Header)
